Broken down flow.scala into different files for better management of code
