+++ src/main/scala/TransformContext.scala
@@ -5,7 +5,7 @@ import reflect.macros.blackbox
 
 abstract class TransformContext[C <: blackbox.Context](override val macroContext: C) extends TransformUtils[C](macroContext) {
 
-  import universe._
+  import macroContext.universe._
 
   val returnName: TermName
   val returnType: Type
